Apprise-Azure/requirements.txt
redis
python-dotenv
apprise
azure-communication-sms
azure-communication-email